Fix card spacing, sorting, layering
For the player's hand and the staging area, card spacing is reduced once
the total number of cards in the zone exceeds a given maximum, to ensure
they can all be displayed in the available space.

Each card has its z-position set based on its index in the list of cards
(equivalent to its x-position) which ensures both neat visual layering
and predictable behaviour when clicking on a card where it overlaps an
adjacent one.

Card movement functions have been extended to allow Vector3 input while
retaining the Vector2 overloads.

Unrelated to this work, several variables have been renamed to be less
generic; in particular the regular use of `cards` to describe almost any
List<Card> reduced readability of the code even when the nature of the
parameter/variable could be inferred from method names etc.