Begin implementing buttons to Activate/Play cards
This commit reinstates the clickable items which allow the player to use
a selected card in different ways.

Currently the implementation is very rough, the clickable items have no
effect but a framework is in place which will facilitate displaying some
defined subset of the options.

Updated design doc with the rationale behind this approach.