Refactor for discarding cards flow
Discard option is now independent of other options that can apply to
selected cards, which better reflects the options that can or cannot be
available at the same time. It is still housed in the options panel for
ease of management.

Added an overload of CardZone.Accept which takes a single Card, so
callers don't have to wrap the Card in a list - this is notably relevant
to the options available for selected cards (Activate/Play/Discard etc).