Partially implement Town to close basic game loop
It is now possible to play repeated rounds, cycling through encounters
followed by the various Town phases which segue into a new encounter.
However, only the Trap encounter has been fully implemented, and much of
the Town functionality is not yet in place.

Since the only available gameplay is to trigger traps until the player
dies, a cursory end-game-and-restart mechanism is also implemented with
this commit. This needs to be revisited and improved on in due course.