Fix incomplete rotation bug
The speed multiplier which is applied to translational card movement was
causing encounter cards to rotate very slowly when returned to the deck
due to the short distance. This particularly affected the agitator card.
Attempts to directly orient the transform (using SetLookRotation), when
the card is killed, were unsuccessful so the variable speed modifier was
removed from rotation mechanics in favour of an appropriate fixed speed.